+++ src/main/java/com/ssafy/questory/repository/QuestRepository.java
@@ -1,7 +1,7 @@
 package com.ssafy.questory.repository;
 
 import com.ssafy.questory.dto.request.quest.QuestRequestDto;
-import com.ssafy.questory.dto.response.quest.AttractionResponseDto;
+import com.ssafy.questory.dto.response.attraction.AttractionResponseDto;
 import com.ssafy.questory.dto.response.quest.QuestsResponseDto;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
@@ -38,6 +38,4 @@ public interface QuestRepository {
     int getQuestCntByQuestId(@Param("questId") int questId);
 
     int getValidQuestCntByQuestId(@Param("questId") int questId);
-
-    List<AttractionResponseDto> searchAttractionByTitle(@Param("searchKeyword") String searchKeyword);
 }
